Update apriori.py
Fixed something about the confidence of a multi items(over 2) frequent set with one-item consequent won't be calculated in function generateRules(). Considerate the situation that a frequent lists L=[[{0},{1},{2}],[{0,1},{1,2},{0,2}],[{0,1,2}],[]], L[2][0]={0,1,2}. The index of this set is 2, which means it will run into the first branch(the rulesFromConseq one) in the original determine statements in line 69(if i>1:). And then in the original function rulesFromConseq(), m=1. So a m+1 items set will be calculated and the confidence of the set with m+1 item consequent was calculated. That is mean the confidence of a multi items(over 2) frequent set with one-item consequent won't be calculated.